requirements.txt
langchain==0.1.20
langgraph==0.0.38
groq==0.30.0
streamlit==1.35.0
pandas==2.2.2
datasets==2.19.1
python-dotenv==1.0.1
